requirements.txt
python-telegram-bot==21.7
yfinance
pandas
httpx
python-dotenv
pytest
pytest-asyncio
fastapi
uvicorn
sqlmodel
asyncpg
alembic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
